xml/parsing/manakai/comments-2.dat

#data
<!DOCTYPE a[
<!---->
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<!--<?y?>-->
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<!---->
<?x?>
<!--]><b/>-->
]><a/>
#errors
1;1;no XML decl;;;s
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
< >
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;bare stago;;;m
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<!  >
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;bogus comment;;;m
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<b>bbb</b>ccc
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;bare stago;;;m
2;4;string in internal subset;;;m
2;7;bare stago;;;m
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<[[xxxx]]>
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;bare stago;;;m
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<!--><?y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;5;bogus comment;;;m
2;6;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?><!--
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
2;10;unclosed comment;;;m
2;10;unclosed internal subset;;;m
2;10;no root element;;;m
#document
| <!DOCTYPE a>
|   <?x ?>

#data
<!DOCTYPE a[
<!---><?y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;6;bogus comment;;;m
2;7;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?><!---
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
2;11;unclosed comment;;;m
2;11;unclosed internal subset;;;m
2;11;no root element;;;m
#document
| <!DOCTYPE a>
|   <?x ?>

#data
<!DOCTYPE a[
<?x?><!--z
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
2;11;unclosed comment;;;m
2;11;unclosed internal subset;;;m
2;11;no root element;;;m
#document
| <!DOCTYPE a>
|   <?x ?>

#data
<!DOCTYPE a[
<?x?><!--z-
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
2;12;unclosed comment;;;m
2;12;unclosed internal subset;;;m
2;12;no root element;;;m
#document
| <!DOCTYPE a>
|   <?x ?>

#data
<!DOCTYPE a[
<!--z--><?y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;9;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?><!--z--
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
2;13;unclosed comment;;;m
2;13;unclosed internal subset;;;m
2;13;no root element;;;m
#document
| <!DOCTYPE a>
|   <?x ?>
